completed day 6 and day 5 (part 1)

+++ 2023/aocutils/StringExtensions.kt
@@ -1,9 +1,9 @@
 package aocutils
 
-fun String.extractAllIntegers(): List<Int> {
+fun String.extractAllIntegers(): List<Long> {
     val regex = Regex("-?\\b\\d+\\b")
     return regex
         .findAll(this)
-        .map { it.value.toInt() }
+        .map { it.value.toLong() }
         .toList()
 }
